refactor searchMovies
