+++ src/hooks/useFetchMovies.js
@@ -1,16 +1,16 @@
 import { useEffect, useState } from 'react';
 import { useLazyFetchMoviesQuery } from '../store/moviesPaginationAPI';
-const useFetchMovies = (page) => {
+const useFetchMovies = ({ page, searchQuery }) => {
   const [isExecuted, setIsExecuted] = useState(false);
   const [fetchMovies] = useLazyFetchMoviesQuery();
 
   useEffect(() => {
     if(!isExecuted) {
-      fetchMovies({ page });
+      fetchMovies({ page, searchQuery });
       setIsExecuted(true);
     }
 
-  }, [page, isExecuted]);
+  }, [page, searchQuery,isExecuted]);
 
   return { fetchMovies }
 }
